+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added `autofit` argument to `axis_geo` which, when enabled, automatically adjusts the sizes of individual interval labels to fit within the interval boxes (#21)
 * Added `title` argument to `axis_geo` which can be used to add a title to the axis with the timescale (#133)
 * Increased efficiency and reduced hackiness of `group_apply` (#135)
+* Downloads related to the palaeorotate function are now retried if they fail (#137)
 
 # palaeoverse 1.4.0